fixed the readme info about py3

+++ defectFinderPeakFinder.py
@@ -273,7 +273,7 @@ def FindPeaks(Hist,outdir,fitdir,canvas,bolVerb = 0):
     intPipeNum = -1
 
   DefectInfo = []
-  for i in range(nMflaws):
+  for i in range(int(nMflaws)):
     DefectInfo = np.append(DefectInfo,i)                                  #Flaw Number
     DefectInfo = np.append(DefectInfo,intPipeNum)                         #Flaw Pipe
     DefectInfo = np.append(DefectInfo,float(MFlaws[i*5+1]))               #Flaw Center
@@ -295,7 +295,7 @@ def GetDefectBoxes(DefectInfo,Ymin,Ymax,bolTempHot = True):
   if bolTempHot == False:
     TScale = 2.
  #Major Flaw Boxes
-  for i in range(len(DefectInfo)/7):
+  for i in range(int(len(DefectInfo)/7)):
     Center = DefectInfo[i*7+2]
     Xmin = Center - 0.5*DefectInfo[i*7+3]
     Xmax = Center + 0.5*DefectInfo[i*7+3] 
@@ -599,7 +599,7 @@ def PrintDefectInfo(DefectInfo,intLine=-1,bolTempHot=True):
   print(" Name    Center     Width     Height    Goodness    Fit")
   nThings = 7
   if bolPrintAll == True:
-    for i in range(len(DefectInfo)/nThings):
+    for i in range(int(len(DefectInfo)/nThings)):
       print("array index: {}".format(int(i*nThings+4)))
       print(DefectInfo)
       SigDefect(DefectInfo[int(i*nThings+4)],bolTempHot)
